[refactor] 채팅기록 불러오기 hook 분리
