+++ src/domains/recommend/components/ChatList.tsx
@@ -1,17 +1,20 @@
 import { useChatScroll } from '../hook/useChatScroll';
-import { ChatListProps } from '../types/recommend';
+import { ChatListProps, RecommendationItem, StepRecommendation } from '../types/recommend';
 import BotMessage from './bot/BotMessage';
 import NewMessageAlert from './bot/NewMessageAlert';
 import UserMessage from './user/UserMessage';
 
-function ChatList({
-  messages,
-  userCurrentStep,
-  onSelectedOption,
-  getRecommendations,
-}: ChatListProps) {
+function ChatList({ messages, userCurrentStep, onSelectedOption }: ChatListProps) {
   const { chatListRef, chatEndRef, showNewMessageAlert, handleCheckBottom, handleScrollToBottom } =
     useChatScroll(messages[messages.length - 1]?.id);
+
+  const getRecommendations = (
+    type: string | undefined,
+    stepData?: StepRecommendation | null
+  ): RecommendationItem[] => {
+    if (type !== 'CARD_LIST' || !stepData?.recommendations) return [];
+    return stepData.recommendations;
+  };
 
   return (
     <div
@@ -32,6 +35,8 @@ function ChatList({
 
           const isTyping = msg.type === 'TYPING';
 
+          const recommendations = getRecommendations(msg.type, msg.stepData);
+
           return (
             <BotMessage
               key={keyId}
@@ -41,7 +46,7 @@ function ChatList({
                   message: msg.message,
                   type: msg.type ?? 'TEXT',
                   options: msg.type === 'RADIO_OPTIONS' ? (msg.stepData?.options ?? []) : [],
-                  recommendations: getRecommendations(msg.type, msg.stepData),
+                  recommendations,
                 },
               ]}
               showProfile={showProfile}
